Array Class

Instantiate an array from an initialized store. Reference: https://github.com/zarr-developers/zarr-python/blob/5dd4a0/zarr/core.py#L51

Details

The Zarr Array class.

Public fields

store

Array store, already initialized.

chunk_store

Separate storage for chunks. If not provided, store will be used for storage of both chunks and metadata.

path

Storage path. String, optional.

read_only

True if array should be protected against modification.

synchronizer

Array synchronizer. Object, optional.

cache_metadata

If True (default), array configuration metadata will be cached. If False, metadata will be reloaded prior to all data access and modification.

cache_attrs

If True (default), user attributes will be cached. If False, attributes will be reloaded prior to all data access and modification.

write_empty_chunks

If True, all chunks will be stored regardless of their contents. If False (default), each chunk is compared to the array's fill value prior to storing. If a chunk is uniformly equal to the fill value, then that chunk is not be stored, and the store entry for that chunk's key is deleted.

Methods

Public methods

Array$new()

Array$get_store()

Array$get_path()

Array$get_name()

Array$get_basename()

Array$get_read_only()

Array$set_read_only()

Array$get_chunk_store()

Array$get_shape()

Array$set_shape()

Array$resize()

Array$get_chunks()

Array$get_dtype()

Array$get_compressor()

Array$get_fill_value()

Array$set_fill_value()

Array$get_order()

Array$get_filters()

Array$get_synchronizer()

Array$get_attrs()

Array$get_ndim()

Array$get_size()

Array$get_itemsize()

Array$get_nbytes()

Array$get_nbytes_stored()

Array$get_cdata_shape()

Array$get_nchunks()

Array$get_nchunks_initialized()

Array$get_is_view()

Array$get_oindex()

Array$get_vindex()

Array$get_write_empty_chunks()

Array$equals()

Array$islice()

Array$length()

Array$get_item()

Array$get_basic_selection()

Array$get_orthogonal_selection()

Array$get_coordinate_selection()

Array$get_mask_selection()

Array$set_item()

Array$set_basic_selection()

Array$set_orthogonal_selection()

Array$set_coordinate_selection()

Array$set_mask_selection()

Array$get_info()

Array$get_digest()

Array$get_hexdigest()

Array$append()

Array$view()

Array$astype()

Array$clone()

Method new()

(Re)load metadata from store.

Create a new Array instance.

Usage

Array$new(
  store,
  path = NA,
  read_only = FALSE,
  chunk_store = NA,
  synchronizer = NA,
  cache_metadata = TRUE,
  cache_attrs = TRUE,
  write_empty_chunks = TRUE
)

Arguments

store

Array store, already initialized.

Returns

An Array instance.

Method get_store()

Usage

Array$get_store()

Method get_path()

Usage

Array$get_path()

Method get_name()

Usage

Array$get_name()

Method get_basename()

Usage

Array$get_basename()

Method get_read_only()

Usage

Array$get_read_only()

Method set_read_only()

Usage

Array$set_read_only(val)

Method get_chunk_store()

Usage

Array$get_chunk_store()

Method get_shape()

Usage

Array$get_shape()

Method set_shape()

Usage

Array$set_shape(value)

Method resize()

Change the shape of the array by growing or shrinking one or more dimensions.

Usage

Array$resize(...)

Method get_chunks()

Usage

Array$get_chunks()

Method get_dtype()

Usage

Array$get_dtype()

Method get_compressor()

Usage

Array$get_compressor()

Method get_fill_value()

Usage

Array$get_fill_value()

Method set_fill_value()

Usage

Array$set_fill_value(val)

Method get_order()

Usage

Array$get_order()

Method get_filters()

Usage

Array$get_filters()

Method get_synchronizer()

Usage

Array$get_synchronizer()

Method get_attrs()

Usage

Array$get_attrs()

Method get_ndim()

Usage

Array$get_ndim()

Method get_size()

Usage

Array$get_size()

Method get_itemsize()

Usage

Array$get_itemsize()

Method get_nbytes()

Usage

Array$get_nbytes()

Method get_nbytes_stored()

Usage

Array$get_nbytes_stored()

Method get_cdata_shape()

Usage

Array$get_cdata_shape()

Method get_nchunks()

Usage

Array$get_nchunks()

Method get_nchunks_initialized()

Usage

Array$get_nchunks_initialized()

Method get_is_view()

Usage

Array$get_is_view()

Method get_oindex()

Usage

Array$get_oindex()

Method get_vindex()

Usage

Array$get_vindex()

Method get_write_empty_chunks()

Usage

Array$get_write_empty_chunks()

Method equals()

Usage

Array$equals(other)

Method islice()

Usage

Array$islice(start = NA, end = NA)

Method length()

Usage

Array$length()

Method get_item()

Usage

Array$get_item(selection)

Method get_basic_selection()

Usage

Array$get_basic_selection(selection = NA, out = NA, fields = NA)

Method get_orthogonal_selection()

Usage

Array$get_orthogonal_selection(selection = NA, out = NA, fields = NA)

Method get_coordinate_selection()

Usage

Array$get_coordinate_selection(selection = NA, out = NA, fields = NA)

Method get_mask_selection()

Usage

Array$get_mask_selection(selection = NA, out = NA, fields = NA)

Method set_item()

Usage

Array$set_item(selection, value)

Method set_basic_selection()

Usage

Array$set_basic_selection(selection, value, fields = NA)

Method set_orthogonal_selection()

Usage

Array$set_orthogonal_selection(selection, value, fields = NA)

Method set_coordinate_selection()

Usage

Array$set_coordinate_selection(selection, value, fields = NA)

Method set_mask_selection()

Usage

Array$set_mask_selection(selection, value, fields = NA)

Method get_info()

Usage

Array$get_info()

Method get_digest()

Usage

Array$get_digest(hashname = "sha1")

Method get_hexdigest()

Usage

Array$get_hexdigest(hashname = "sha1")

Method append()

Usage

Array$append(data, axis = 0)

Method view()

Usage

Array$view(
  shape = NA,
  chunks = NA,
  dtype = NA,
  fill_value = NA,
  filters = NA,
  read_only = NA,
  synchronizer = NA
)

Method astype()

Usage

Array$astype(dtype)

Method clone()

The objects of this class are cloneable with this method.

Usage

Array$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.