Attributes Class

Class providing access to user attributes on an array or group. Reference: https://github.com/zarr-developers/zarr-python/blob/5dd4a0/zarr/attrs.py#L7

Details

The Zarr Attributes class.

Public fields

cached_aslist

The attributes cached as a list.

store

Attributes store, already initialized.

key

The key under which the attributes will be stored.

read_only

If True, attributes cannot be modified.

cache

If True (default), attributes will be cached locally.

synchronizer

Only necessary if attributes may be modified from multiple threads or processes.

Active bindings

cached_aslist

The attributes cached as a list.

Methods

Public methods

Attributes$new()

Attributes$asdict()

Attributes$refresh()

Attributes$contains()

Attributes$get_item()

Attributes$set_item()

Attributes$clone()

Method new()

Create a new Attributes instance.

Usage

Attributes$new(
  store,
  key = ".zattrs",
  read_only = FALSE,
  cache = TRUE,
  synchronizer = NA
)

Arguments

store

Attributes store, already initialized.

Returns

An Attributes instance.

Method asdict()

Usage

Attributes$asdict()

Method refresh()

Usage

Attributes$refresh()

Method contains()

Usage

Attributes$contains(x)

Method get_item()

Usage

Attributes$get_item(item)

Method set_item()

Usage

Attributes$set_item(item, value)

Method clone()

The objects of this class are cloneable with this method.

Usage

Attributes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.