MemoryStore Class

Store class that uses a hierarchy of list objects, thus all data will be held in main memory. Reference: https://github.com/zarr-developers/zarr-python/blob/a5dfc3b4/zarr/storage.py#L512

Details

MemoryStore for Zarr

Super class

pizzarr::Store -> MemoryStore

Public fields

root

The root list for the store.

Methods

Public methods

MemoryStore$new()

MemoryStore$get_parent()

MemoryStore$get_item()

MemoryStore$set_item()

MemoryStore$contains_item()

MemoryStore$clone()

Inherited methods

Method new()

Create a new memory store.

Usage

MemoryStore$new()

Returns

A new MemoryStore object.

Method get_parent()

Get the parent of an item.

Usage

MemoryStore$get_parent(item)

Arguments

item

The item key.

Returns

A list with the keys parent and key.

Method get_item()

Get an item from the store.

Usage

MemoryStore$get_item(item)

Arguments

item

The item key.

Returns

The item data in a vector of type raw.

Method set_item()

Set an item in the store.

Usage

MemoryStore$set_item(item, value)

Arguments

item

The item key.

value

The item value as a vector of type raw.

Method contains_item()

Determine whether the store contains an item.

Usage

MemoryStore$contains_item(item)

Arguments

item

The item key.

Returns

A boolean value.

Method clone()

The objects of this class are cloneable with this method.

Usage

MemoryStore$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.